# chore: speed up local dev