--- a/package.json
+++ b/package.json
@@ -15,7 +15,7 @@
     "build": "NODE_ENV=production webpack --config scripts/webpack/webpack.prod.ts",
     "build:panel": "NODE_ENV=production webpack --config scripts/webpack/webpack.panel.ts",
     "build:size-limit": "NODE_ENV=production webpack --config scripts/webpack/webpack.size-limit.ts",
-    "test": "jest --maxWorkers=1",
+    "test": "jest --runInBand",
     "test:ss": "UPDATE_SNAPSHOTS=true ./scripts/jest-snapshots/run-docker.sh",
     "test:ss-check": "./scripts/jest-snapshots/run-docker.sh",
     "lint": "eslint ./ --ext .js,.jsx,.ts,.tsx --cache --fix",