--- a/scripts/webpack/shared.ts
+++ b/scripts/webpack/shared.ts
@@ -61,35 +61,42 @@ export function getJsLoader() {
       exclude: /node_modules/,
       use: [
         {
-          loader: 'babel-loader',
+          loader: 'esbuild-loader',
           options: {
-            cacheDirectory: true,
-            babelrc: true,
-
-            plugins: ['@babel/plugin-transform-runtime'],
-            // Note: order is bottom-to-top and/or right-to-left
-            presets: [
-              [
-                '@babel/preset-env',
-                {
-                  targets: {
-                    browsers: 'last 3 versions',
-                  },
-                  useBuiltIns: 'entry',
-                  corejs: 3,
-                  modules: false,
-                },
-              ],
-              [
-                '@babel/preset-typescript',
-                {
-                  allowNamespaces: true,
-                },
-              ],
-              '@babel/preset-react',
-            ],
+            loader: 'tsx', // Or 'ts' if you don't need tsx
+            target: 'es2015',
           },
         },
+        //        {
+        //          loader: 'babel-loader',
+        //          options: {
+        //            cacheDirectory: true,
+        //            babelrc: true,
+        //
+        //            plugins: ['@babel/plugin-transform-runtime'],
+        //            // Note: order is bottom-to-top and/or right-to-left
+        //            presets: [
+        //              [
+        //                '@babel/preset-env',
+        //                {
+        //                  targets: {
+        //                    browsers: 'last 3 versions',
+        //                  },
+        //                  useBuiltIns: 'entry',
+        //                  corejs: 3,
+        //                  modules: false,
+        //                },
+        //              ],
+        //              [
+        //                '@babel/preset-typescript',
+        //                {
+        //                  allowNamespaces: true,
+        //                },
+        //              ],
+        //              '@babel/preset-react',
+        //            ],
+        //          },
+        //        },
       ],
     },
   ];