--- a/tsconfig.eslint.json
+++ b/tsconfig.eslint.json
@@ -1,4 +1,7 @@
 {
   "extends": "./tsconfig.json",
-  "include": ["."]
+  "include": ["."],
+  // Since we excluded the test files in the main tsconfig.json
+  // We must renable them by explicitly setting an 'exclude'
+  "exclude": ["node_modules"]
 }