--- a/webapp/README.md
+++ b/webapp/README.md
@@ -38,3 +38,12 @@ To be able to do that, you need to add the alias to the following files:
 * `scripts/webpack/webpack.common.js`
 * `tsconfig.json`
 * `jest.config.js`
+
+# Developing the webapp/templates page
+By default, developing pages other than the index require a bit of setup:
+
+
+For example, acessing http://locahlost:4040/forbidden won't work
+To be able to access it, update the variable `pages` in `scripts/webpack.common.ts` to allow building all pages when in dev mode.
+
+Beware, this will make the (local) build slower.