* `maxWorkers=1` as per https://github.com/kulshekhar/ts-jest/issues/1044
* a scoped `includes` to only care about the webapp
* a `tsconfig` for tests where we manually set which `types` are required
* different eslint rules for tests
* use `esbuild` instead of babel
* use `webpack-livereload-plugin` to add a livereload server to the html files served by go (removing the need for the double server shenanigans)
* in local dev only build the `index.html` since that's the most common case (see webapp/README.md for more info)

---
# Local build
Before
```
21:32:05 webpack | 0 (webpack 5.65.0) compiled with 1 warning in 3296 ms
```

After
```
21:33:46 webpack | webpack 5.65.0 compiled with 1 warning in 446 ms
```


# Tests
Before
```
Test Suites: 15 passed, 15 total
Tests:       193 passed, 193 total
Snapshots:   4 passed, 4 total
Time:        25.552 s
Ran all test suites.
Done in 26.96s.
```

After
```
Test Suites: 15 passed, 15 total
Tests:       193 passed, 193 total
Snapshots:   4 passed, 4 total
Time:        17.276 s
Ran all test suites.
Done in 18.88s.
```

But the biggest improvement seems to be when running tests with `--watch`. 

Before
```
Test Suites: 1 failed, 1 total
Tests:       3 failed, 3 passed, 6 total
Snapshots:   0 total
Time:        8.744 s, estimated 10 s
Ran all test suites matching /Sidebar.spec.tsx/i.
```

After
```
Test Suites: 1 failed, 1 total
Tests:       3 failed, 3 passed, 6 total
Snapshots:   0 total
Time:        1.753 s, estimated 2 s
Ran all test suites matching /Sidebar.spec.tsx/i.
```

# Prod Build

Before
```
$ time yarn build 
(...) 
Done in 21.36s.

________________________________________________________
Executed in   21.53 secs    fish           external
   usr time   38.73 secs    0.51 millis   38.73 secs
   sys time    0.95 secs    1.05 millis    0.95 secs

````

After
```
$ time yarn build  
(...)
Done in 11.45s.

________________________________________________________
Executed in   11.61 secs    fish           external
   usr time   19.26 secs  811.00 micros   19.26 secs
   sys time    0.66 secs  909.00 micros    0.66 secs
```