@petethepig 2022-01-08
@eh-am `maxWorkers` one is quite unexpected