@eh-am 2022-01-08
One thing I noticed is that disabling `WebpackPluginServe` makes the build much much faster, but then you need to manually refresh the page

`WebpackPluginServe` turned on
```
0 (webpack 5.65.0) compiled with 1 warning in 3533 ms
```

`WebpackPluginServe` turned off
```
webpack 5.65.0 compiled with 1 warning in 305 ms
```